refactor: AI correction compat 검증 로직 서비스 분리 (#297)
- controller에 있던 compat 입력 검증과 결과 정규화를 application service로 이동합니다.
- internal controller는 thin controller 역할만 유지하도록 계층 책임을 정리합니다.
